+++ SearchThing/src/Extensions/Panel/Filter/PropTagSearchPanel.cs
@@ -6,8 +6,8 @@ namespace SearchThing.Extensions.Panel.Filter;
 public class PropTagSearchPanel : FilterSearchSearchPanel
 {
     public override string Tag => "Props";
-    protected override bool Filter(SearchableCrate searchableCrate)
+    protected override bool Filter(ISearchableCrate searchableCrate)
     {
-        return searchableCrate.CrateType == CrateType.Prop;
+        return searchableCrate is { CrateType: CrateType.Prop, Redacted: false };
     }
 }