+++ SearchThing/src/Extensions/Panel/History/AvatarHistorySearchPanel.cs
@@ -7,7 +7,7 @@ public class AvatarHistorySearchPanel : HistorySearchPanel
 {
 
     public override string Tag => "Avatar History";
-    protected override bool Filter(HistoryEntry entry)
+    protected override bool Filter(ISearchableCrate entry)
     {
         return entry.CrateType == CrateType.Avatar;
     }