+++ SearchThing/src/Mod.cs
@@ -1,10 +1,9 @@
 ﻿using BoneLib;
 using SearchThing;
 using MelonLoader;
-using SearchThing.Lookup.Cache;
 using SearchThing.Presets;
 
-[assembly: MelonInfo(typeof(Mod), "SearchThing", "0.4.0", "Mash")]
+[assembly: MelonInfo(typeof(Mod), "SearchThing", "0.4.1", "Mash")]
 [assembly: MelonGame("Stress Level Zero", "BONELAB")]
 namespace SearchThing;
 
@@ -22,11 +21,13 @@ public class Mod : MelonMod
     private static void OnWarehouseReady()
     {
         DatabaseManager.OnMelonInitialize();
+        SearchManager.InitializeSearchThread();
         PresetManager.LoadPresets();
     }
 
     public override void OnDeinitializeMelon()
     {
         DatabaseManager.OnMelonDeinitialize();
+        SearchManager.ShutdownSearchThread();
     }
 }